+++ CHANGELOG.md
@@ -9,3 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - plugin init
+
+### Changed
+- plugin config dumping
+- plugin datasource binding